+++ src/modules/scenegraph/SceneGraphUtil.cpp
@@ -39,6 +39,7 @@ static void copy(const SceneGraphNode &node, SceneGraphNode &target, bool copyKe
 	}
 	target.setVisible(node.visible());
 	target.setLocked(node.locked());
+	target.setOpacity(node.opacity());
 	target.setPivot(node.pivot());
 	target.setColor(node.color());
 	target.addProperties(node.properties());